+++ cxx/gewurz/bidiag.cuh
@@ -7,8 +7,8 @@
 
 namespace gw {
 
-auto CuDot(thrust::device_vector<CuCxH> const &a, thrust::device_vector<CuCxH> const &b) -> __half;
-auto CuNorm(thrust::device_vector<CuCxH> const &a) -> __half;
+auto CuDot(thrust::device_vector<CuCxH> const &a, thrust::device_vector<CuCxH> const &b) -> __nv_bfloat16;
+auto CuNorm(thrust::device_vector<CuCxH> const &a) -> __nv_bfloat16;
 
 template <typename T, typename U> void CuScale(thrust::device_vector<T> &a, U const b)
 {
@@ -54,56 +54,56 @@ template <typename T, int xRank, int yRank> struct Bidiag
     if (Minv) {
       thrust::copy(b.vec.begin(), b.vec.end(), Mu.vec.begin());
       Minv->forward(Mu.span, u.span);
-      β = __half2float(cuda::std::sqrt(CuDot(Mu.vec, u.vec)));
-      CuScale(Mu.vec, __float2half(1.f / β));
+      β = __bfloat162float(cuda::std::sqrt(CuDot(Mu.vec, u.vec)));
+      CuScale(Mu.vec, __float2bfloat16(1.f / β));
     } else {
       thrust::copy(b.vec.begin(), b.vec.end(), u.vec.begin());
-      β = __half2float(cuda::std::sqrt(CuDot(Mu.vec, u.vec)));
+      β = __bfloat162float(cuda::std::sqrt(CuDot(Mu.vec, u.vec)));
     }
-    CuScale(u.vec, __float2half(1.f / β));
+    CuScale(u.vec, __float2bfloat16(1.f / β));
 
     if (Ninv) {
       A->adjoint(u.span, Nv.span);
       Ninv->forward(Nv.span, v.span);
-      α = __half2float(cuda::std::sqrt(CuDot(Nv.vec, v.vec)));
-      CuScale(Nv.vec, __float2half(1 / α));
+      α = __bfloat162float(cuda::std::sqrt(CuDot(Nv.vec, v.vec)));
+      CuScale(Nv.vec, __float2bfloat16(1 / α));
     } else {
       A->adjoint(u.span, v.span);
-      α = __half2float(cuda::std::sqrt(CuDot(v.vec, v.vec)));
+      α = __bfloat162float(cuda::std::sqrt(CuDot(v.vec, v.vec)));
     }
-    CuScale(v.vec, __float2half(1.f / α));
+    CuScale(v.vec, __float2bfloat16(1.f / α));
   }
 
   void next()
   {
     if (Minv) {
       A->forward(v.span, u.span);
-      CuAsubBC2B(u.vec, Mu.vec, __float2half(α));
+      CuAsubBC2B(u.vec, Mu.vec, __float2bfloat16(α));
       Minv->forward(Mu.span, u.span);
-      β = __half2float(cuda::std::sqrt(CuDot(Mu.vec, u.vec)));
-      CuScale(Mu.vec, __float2half(1.f / β));
+      β = __bfloat162float(cuda::std::sqrt(CuDot(Mu.vec, u.vec)));
+      CuScale(Mu.vec, __float2bfloat16(1.f / β));
     } else {
       A->forward(v.span, Mu.span);
-      CuAsubBC2B(Mu.vec, u.vec, __float2half(α));
-      β = __half2float(cuda::std::sqrt(CuDot(u.vec, u.vec)));
+      CuAsubBC2B(Mu.vec, u.vec, __float2bfloat16(α));
+      β = __bfloat162float(cuda::std::sqrt(CuDot(u.vec, u.vec)));
     }
-    CuScale(u.vec, __float2half(1.f / β));
+    CuScale(u.vec, __float2bfloat16(1.f / β));
 
     if (Ninv) {
       A->adjoint(u.span, v.span);
-      CuAsubBC2B(v.vec, Nv.vec, __float2half(β));
+      CuAsubBC2B(v.vec, Nv.vec, __float2bfloat16(β));
       Ninv->forward(Nv.span, v.span);
-      α = __half2float(cuda::std::sqrt(CuDot(Nv.vec, v.vec)));
-      CuScale(Nv.vec, __float2half(1.f / α));
+      α = __bfloat162float(cuda::std::sqrt(CuDot(Nv.vec, v.vec)));
+      CuScale(Nv.vec, __float2bfloat16(1.f / α));
     } else {
-      fmt::print(stderr, "Before |u| {} |v| {}\n", __half2float(CuNorm(u.vec)), __half2float(CuNorm(v.vec)));
+      fmt::print(stderr, "Before |u| {} |v| {}\n", __bfloat162float(CuNorm(u.vec)), __bfloat162float(CuNorm(v.vec)));
       A->adjoint(u.span, Nv.span);
-      fmt::print(stderr, "Middle |u| {} |Nv| {}\n", __half2float(CuNorm(u.vec)), __half2float(CuNorm(Nv.vec)));
-      CuAsubBC2B(Nv.vec, v.vec, __float2half(β));
-      fmt::print(stderr, "After |Nv| {} |v| {}  β {}\n", __half2float(CuNorm(u.vec)), __half2float(CuNorm(v.vec)), β);
-      α = __half2float(cuda::std::sqrt(CuDot(v.vec, v.vec)));
+      fmt::print(stderr, "Middle |u| {} |Nv| {}\n", __bfloat162float(CuNorm(u.vec)), __bfloat162float(CuNorm(Nv.vec)));
+      CuAsubBC2B(Nv.vec, v.vec, __float2bfloat16(β));
+      fmt::print(stderr, "After |Nv| {} |v| {}  β {}\n", __bfloat162float(CuNorm(u.vec)), __bfloat162float(CuNorm(v.vec)), β);
+      α = __bfloat162float(cuda::std::sqrt(CuDot(v.vec, v.vec)));
     }
-    CuScale(v.vec, __float2half(1.f / α));
+    CuScale(v.vec, __float2bfloat16(1.f / α));
   }
 };
 } // namespace gw
